feat(cdk:forms): support disabled status
refactor(cdk:forms): resolve circular dependencies
