+++ packages/cdk/forms/__tests__/formControl.spec.ts
@@ -1,5 +1,5 @@
 import { flushPromises } from '@vue/test-utils'
-import { FormControl } from '../src/controls/formControl'
+import { FormControl } from '../src/controls'
 import { ValidateErrors } from '../src/types'
 import { Validators } from '../src/validators'
 
@@ -133,6 +133,36 @@ describe('formControl.ts', () => {
 
       await control.validate()
 
+      expect(control.hasError('required')).toEqual(false)
+    })
+  })
+
+  describe('disabled work', () => {
+    let control: FormControl<string>
+
+    test('default disabled work', async () => {
+      control = new FormControl('', { validators: Validators.required, disabled: true })
+
+      expect(control.disabled.value).toEqual(true)
+      expect(control.status.value).toEqual('valid')
+      expect(control.hasError('required')).toEqual(false)
+    })
+
+    test('disable and enable work', async () => {
+      control = new FormControl('', { validators: Validators.required, disabled: true })
+
+      control.enable()
+      await flushPromises()
+
+      expect(control.disabled.value).toEqual(false)
+      expect(control.status.value).toEqual('invalid')
+      expect(control.hasError('required')).toEqual(true)
+
+      control.disable()
+      await flushPromises()
+
+      expect(control.disabled.value).toEqual(true)
+      expect(control.status.value).toEqual('valid')
       expect(control.hasError('required')).toEqual(false)
     })
   })
