+++ packages/cdk/forms/index.ts
@@ -1,8 +1,4 @@
-export * from './src/controls/abstractControl'
-export * from './src/controls/formArray'
-export * from './src/controls/formControl'
-export * from './src/controls/formGroup'
-export * from './src/controls/types'
+export * from './src/controls'
 export * from './src/typeof'
 export * from './src/types'
 export * from './src/useForms'
