+++ packages/cdk/forms/src/typeof.ts
@@ -1,25 +1,17 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import type { FormArray } from './controls/formArray'
-import type { FormGroup } from './controls/formGroup'
-import type { FormControl } from './controls/formControl'
-
-import { hasOwnProperty, isArray } from '@idux/cdk/utils'
-import { AbstractControl } from './controls/abstractControl'
+import { AbstractControl, FormControl, FormGroup, FormArray } from './controls'
 
 export const isAbstractControl = (val: unknown): val is AbstractControl => {
   return val instanceof AbstractControl
 }
 
-// Since AbstractControl be dependent on the function, `val instanceof FormArray` cannot be used here.
-export const isFormArray = (val: unknown): val is FormArray => {
-  return isAbstractControl(val) && !isFormControl(val) && isArray((val as FormArray).controls.value)
-}
-
-// Since AbstractControl be dependent on the function, `val instanceof FormGroup` cannot be used here.
-export const isFormGroup = (val: unknown): val is FormGroup => {
-  return isAbstractControl(val) && !isFormControl(val) && !isFormArray(val)
-}
-
 export const isFormControl = (val: unknown): val is FormControl => {
-  return isAbstractControl(val) && hasOwnProperty(val, '_initValue')
+  return val instanceof FormControl
+}
+
+export const isFormGroup = (val: unknown): val is FormGroup => {
+  return val instanceof FormGroup
+}
+
+export const isFormArray = (val: unknown): val is FormArray => {
+  return val instanceof FormArray
 }
