+++ packages/cdk/forms/src/useValueAccessor.ts
@@ -1,21 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type { AbstractControl } from './controls/abstractControl'
+import type { AbstractControl } from './controls'
 
 import { reactive, getCurrentInstance, watchEffect, computed, ComputedRef } from 'vue'
 import { isNil, Logger } from '@idux/cdk/utils'
 import { isAbstractControl } from './typeof'
 import { injectControl, injectControlOrPath } from './utils'
-
-interface AccessorOptions {
-  controlKey?: string
-  valueKey?: string
-}
-
-export interface ValueAccessor {
-  value: any
-  setValue: (value: any) => void
-  markAsBlurred: () => void
-}
 
 export function useValueControl(controlKey = 'control'): ComputedRef<AbstractControl | null> {
   const { props } = getCurrentInstance()!
@@ -37,8 +26,21 @@ export function useValueControl(controlKey = 'control'): ComputedRef<AbstractCon
   })
 }
 
+export interface AccessorOptions {
+  controlKey?: string
+  valueKey?: string
+  disabledKey?: string
+}
+
+export interface ValueAccessor {
+  value: any
+  setValue: (value: any) => void
+  markAsBlurred: () => void
+  disabled: boolean
+}
+
 export function useValueAccessor(options: AccessorOptions = {}): ValueAccessor {
-  const { controlKey, valueKey = 'value' } = options
+  const { controlKey, valueKey = 'value', disabledKey = 'disabled' } = options
   const accessor = reactive<ValueAccessor>({} as ValueAccessor)
   const { props, emit } = getCurrentInstance()!
 
@@ -50,6 +52,7 @@ export function useValueAccessor(options: AccessorOptions = {}): ValueAccessor {
       accessor.value = currControl.valueRef
       accessor.setValue = (value: any) => currControl.setValue(value, { dirty: true })
       accessor.markAsBlurred = () => currControl.markAsBlurred()
+      accessor.disabled = currControl.disabled as unknown as boolean
     } else {
       accessor.setValue = value => {
         accessor.value = value
@@ -62,8 +65,8 @@ export function useValueAccessor(options: AccessorOptions = {}): ValueAccessor {
   watchEffect(() => {
     if (!control.value) {
       accessor.value = props[valueKey]
+      accessor.disabled = props[disabledKey] as boolean
     }
   })
-
   return accessor
 }
