+++ packages/components/form/demo/Coordinated.md
@@ -7,12 +7,16 @@ order: 3
 
 ## zh
 
-使用 `setValue` 来动态设置其他控件的值。
+使用 `disable` 和 `enable` 来禁用或者启用控件。
 
-使用 `setValidators` 来动态设置其他控件的校验规则。
+使用 `setValue` 来动态设置控件的值。
+
+使用 `setValidators` 来动态设置控件的校验规则。
 
 ## en
 
-Use `setValue` to set other control's value according to different situations.
+Use `disable` and `enable` to disable or enable control.
 
-Use `setValidators` to set other control's validators according to different situations.
+Use `setValue` to set control's value according to different situations.
+
+Use `setValidators` to set control's validators according to different situations.
